connect login and reset pwd
