--- app/page.tsx
@@ -1,8 +1,0 @@
-import { Button } from "@/components/ui/button";
-export default function Home() {
-  return (
-    <>
-      <div>HelloWorldA!</div>;<Button>Default</Button>
-    </>
-  );
-}
