+++ app/reset_password/layout.tsx
@@ -1,0 +1,7 @@
+import type { Metadata } from "next"
+import Page from "./page"
+
+export const metadata: Metadata = {
+  title: "Reset Password",
+}
+export default Page
